store len with CDB in journal, start on kv stuff

+++ list_experiment/src/mock_pool.rs
@@ -5,9 +5,9 @@ pub struct MockPool {
 }
 
 impl MockPool {
-    pub fn new(size: usize) -> Self {
+    pub fn new(size: u64) -> Self {
         Self {
-            contents: vec![0; size],
+            contents: vec![0; size as usize],
         }
     }
 }
